form nonces All forms have been updated where neccesary with nonces In the allergen_form the unused forms like the update and other premium features have been removed Some database functions for the table have been removed as well either because it is a premium version or it was deemed unneccesary

+++ allergens-dietary/php/woocommerce/product_settings.php
@@ -77,6 +77,7 @@ class Allergens_Dietary_Product_Settings
 
 		?>
 		<div id="allergens_dietary_ictoria_product_data" class="panel woocommerce_options_panel">
+			<?php wp_nonce_field('allergen-product-action','allergen-product-nonce'); ?>
 		<h2><?php echo esc_html__( 'Select allergen(s) and/or dietary restrictions:', 'allergens-dietary' ); ?></h2>
 		<?php foreach ( $allergens as $allergen ) : ?>
 			<?php 
@@ -120,38 +121,46 @@ class Allergens_Dietary_Product_Settings
 	 */
 	public function save_product_options($post_id)
 	{
-		$allergensSelected = array();
-		$allergenNames = array();
-		unset($this->_attachedAllergens);
-		$this->_attachedAllergens = Allergens_Dietary_Allergy_Product_Queries::getInstance()->getAllergyProduct($post_id);
-
-		foreach ($this->_allergens as $allergen) {
-			if (isset($_POST[($this->replace_space_chars($allergen['allergy_name']) . '_allergens_dietary_ictoria')])) {
-				$allergensSelected[] = $allergen['allergy_name'];
+		if(isset($_POST['allergen-product-nonce']) && !empty($_POST['allergen-product-nonce']))
+		{
+			$nonce = sanitize_text_field(wp_unslash($_POST['allergen-product-nonce']));
+			if(!wp_verify_nonce($nonce, 'allergen-product-action')){
+				wp_die(esc_html(__('Something went wrong!','allergens-dietary')));
 			}
-		}
-		// Set old allergen name settings
-		foreach ($this->_attachedAllergens as $allergen) {
-			$allergenNames[] = $allergen['allergy_name'];
-		}
-		
-		$new_diff = array_diff($allergensSelected, $allergenNames);
-		$old_diff = array_diff($allergenNames, $allergensSelected);
-
-		// If the new doesnt contain allergens from the old one, Delete the old.
-		if ($old_diff) {
-			foreach ($old_diff as $allergen) {
-				Allergens_Dietary_Allergy_Product_Queries::getInstance()->deleteAllergyProduct($post_id, $allergen);
+			
+			$allergensSelected = array();
+			$allergenNames = array();
+			unset($this->_attachedAllergens);
+			$this->_attachedAllergens = Allergens_Dietary_Allergy_Product_Queries::getInstance()->getAllergyProduct($post_id);
+	
+			foreach ($this->_allergens as $allergen) {
+				if (isset($_POST[($this->replace_space_chars($allergen['allergy_name']) . '_allergens_dietary_ictoria')])) {
+					$allergensSelected[] = $allergen['allergy_name'];
+				}
 			}
-		}
-		// If the old doesn't contain allergens from the new one, Add the new.
-		if ($new_diff) {
-			foreach ($new_diff as $allergen) {
-				Allergens_Dietary_Allergy_Product_Queries::getInstance()->addAllergyProduct($post_id, $allergen);
+			// Set old allergen name settings
+			foreach ($this->_attachedAllergens as $allergen) {
+				$allergenNames[] = $allergen['allergy_name'];
 			}
+			
+			$new_diff = array_diff($allergensSelected, $allergenNames);
+			$old_diff = array_diff($allergenNames, $allergensSelected);
+	
+			// If the new doesnt contain allergens from the old one, Delete the old.
+			if ($old_diff) {
+				foreach ($old_diff as $allergen) {
+					Allergens_Dietary_Allergy_Product_Queries::getInstance()->deleteAllergyProduct($post_id, $allergen);
+				}
+			}
+			// If the old doesn't contain allergens from the new one, Add the new.
+			if ($new_diff) {
+				foreach ($new_diff as $allergen) {
+					Allergens_Dietary_Allergy_Product_Queries::getInstance()->addAllergyProduct($post_id, $allergen);
+				}
+			}
+	
+			return;
 		}
-
-		return;
 	}
 
 	private function replace_space_chars(string $allergens): string
